+++ example/main.dart
@@ -18,12 +18,20 @@ class MyApp extends StatelessWidget {
             children: <Widget>[
               Center(
                 child: NumberSelection(
+                  theme: NumberSelectionTheme(
+                      draggableCircleColor: Colors.blue,
+                      iconsColor: Colors.white,
+                      numberColor: Colors.white,
+                      backgroundColor: Colors.deepPurpleAccent,
+                      outOfConstraintsColor: Colors.deepOrange),
                   initialValue: 1,
                   minValue: -10,
                   maxValue: 10,
                   direction: Axis.vertical,
-                  withSpring: false,
+                  withSpring: true,
                   onChanged: (int value) => print("value: $value"),
+                  onOutOfConstraints: () =>
+                      print("This value is too high or too low"),
                 ),
               ),
               Center(
@@ -32,7 +40,7 @@ class MyApp extends StatelessWidget {
                   minValue: -1,
                   maxValue: 10,
                   direction: Axis.horizontal,
-                  withSpring: false,
+                  withSpring: true,
                   onChanged: (int value) => print("value: $value"),
                 ),
               ),
@@ -43,5 +51,3 @@ class MyApp extends StatelessWidget {
     );
   }
 }
-
-