+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## [0.0.1] - TODO: Add release date.
+## [1.0.0] - 05/02/2021
 
-* TODO: Describe initial release.
+* Added NumberSelectionTheme property to customize style of widget
+* Added animation when user try to change value to a number that is outside defined range
+ * Added onOutOfConstraints property to trigger an event when value is outside defined range
+
+## [0.0.1] - Initial Release